I've changed my mind one more time. Model of JOIN nodes that have tokens wating on the nodes before join, until all the tokens will gather, is fragile, makes for complicated code, and I dont like it.
Besides - if we have

(start 1) -> (fork)
(start 2) -> (A)
(A) - [external condition == false] -> (join)
(fork) -> (join)
(fork) -> (A)

Then we can't evaluate this, if we have model of joins like I've described in the begining of this commit message.
Because join stops one part of fork from propagating.

So I revert to conception of join that holds tokens wating inside the node, and just waits until there are as many tokens,
as there is incoming transitions. This is a little naive - if there is something sending tokens by one transition to join - it
will eventually produce enough tokens for join to propagate further. But I don't want tokens to contain history of
preious transitions/edges, and without that it would be hard to decide where token has been.

I'll see if it makes problems.

To avoid other problems - FORK and JOIN nodes allo NO CONDITIONS on OUTGOING TRANSITIONS.

Node can be fork and join at once.